+++ HISTORY.rst
@@ -184,6 +184,7 @@
 * Show the full title as a hover title even when truncated.
 * Increase default title truncation length to 120 characters.
 * Improve inherited permission computation.
+* Add test case mixin that produces ephimeral models.
 
 3.1.11 (2019-04-XX)
 ===================